requirements.txt
click==8.0.1
dataclasses==0.8
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.19.5
Pillow==8.3.2
SQLAlchemy==1.4.23
tqdm==4.62.2
typing-extensions==3.10.0.2
Werkzeug==2.0.1
zipp==3.5.0
